LPS-154737 Move IG display context classes to DM api module
