+++ modules/apps/document-library/document-library-api/src/main/java/com/liferay/document/library/display/context/IGViewFileVersionDisplayContext.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.image.gallery.display.kernel.display.context;
+package com.liferay.document.library.display.context;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.servlet.taglib.ui.Menu;
